CR Update: Move default arrow config to cornerstone instead of default
