+++ extensions/cornerstone/package.json
@@ -47,8 +47,8 @@
   "dependencies": {
     "@babel/runtime": "^7.5.5",
     "classnames": "^2.2.6",
+    "lodash.merge": "^4.6.2",
     "lodash.throttle": "^4.1.1",
-    "lodash.isempty": "^4.1.1",
     "query-string": "^6.8.3",
     "react-cornerstone-viewport": "2.x.x"
   }
